Clear preferences daemon
Finally we are getting our preferences to stick by killing a daemon that
was caching the old preferences and ignoring our changes to the plist.

Also adds setting to close window on clean exit.

Finally cleans up old attempts and killing daemons and files that we
thought were holding on to the cached prefs.