osx_defaults does proper escaping
We now use Shellwords library to properly escape keys and values (if a
string type) of the default being written.  This way the users of this
utility don't have to worry about escaping things themselves.  This
required changing the terminal_preferences to no longer attempt to
excape some of it's keys.